Fixed navigation for Sleep Assessment: -renamed SleepAssessment.tsx to SleepAssessmentScreen.tsx, which was causing naming inconsistencies -changed import of useNavigation from an expo one to react native, thank you stack overflow
